+++ changelog.md
@@ -1,3 +1,15 @@
+# qase-java 4.1.12
+
+## What's new
+
+Added support for configurations in test runs. You can now specify configuration groups and values that will be associated with your test runs. This includes:
+- JSON configuration format with `values` array and `createIfNotExists` flag
+- Environment variable support: `QASE_TESTOPS_RUN_CONFIGURATIONS="browser=chrome,environment=staging"`
+- Environment variable support for createIfNotExists: `QASE_TESTOPS_CONFIGURATIONS_CREATE_IF_NOT_EXISTS="true"`
+- System property support: `-DQASE_TESTOPS_RUN_CONFIGURATIONS="browser=firefox,os=linux"`
+- System property support for createIfNotExists: `-DQASE_TESTOPS_CONFIGURATIONS_CREATE_IF_NOT_EXISTS="true"`
+- Automatic creation of configuration groups and values when `createIfNotExists` is true
+
 # qase-java 4.1.11
 
 ## What's new